Add plotting of electric field and drift saving.
The plotting of the electric field in amplification is optional an can
be controlled within the simulation.conf file (see template for
reference). For saving the drift more advanced parameters are available
right now. You can set the number of skipping points and also the lines
are now correctly associated to one event.