Add event browsing without the need to restart.
To make it comfortable to navigate between events without restarting the
viewer a navigation system is added. It enables switching between the
events by pressing the left/right arrow buttons to go to the next event.
As for now the navigation does not check if the "asked for" event number
exceeds the number of total events.