Add option to change bounding box and grid.
To make visualisation more clear an option is added to change the size
of the bounding box via a command line parameter 'bounding-box' which
takes a list of tuples as a string. To make it easier to see the size of
the grid every 0.5 cm a line is drawn in orange.