First version of the new run manager.
The run manager should make it easier to setup different configurations
and compute the results on MOGON. It will compile the current source
when a new job is created and copies all necessary file to a dedicated
run directory. While doing this it will also create the output
directory and copies (if asked to) the particleconversion input file
there.