Add drift info to output and correct e-counts.
To make gain analysis easier store the number of electrons generated in
the drift simulation also into the output file of the avalanche. That
way only one file has to be read in for the analysis. Also the number of
generated electrons and electron end points per event was not correctly
stored before which is corrected here.