Allow saving the simulation before job times out.
To get data for runs that are to long to finish in time, the job script
as well as the actual simulation program now ends gracefully on receiving
a SIGUSR2 signal. The job submission is configured in a way, that this
signal is send 10 minutes before the job reaches its time limit.